rag_flow/pyproject.toml
[project]
name = "rag_flow"
version = "0.1.0"
description = "rag_flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "azure-ai-textanalytics>=5.3.0",
    "beautifulsoup4>=4.13.5",
    "crewai[tools]>=0.193.2,<1.0.0",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.33",
    "markdown>=3.9",
    "numpy>=2.2.6",
    "numpydoc>=1.9.0",
    "openai>=1.109.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "pymupdf>=1.26.4",
    "pypdf>=6.1.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "ragas>=0.3.5",
    "requests>=2.32.5",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
    "streamlit>=1.50.0",
    "tiktoken>=0.11.0",
    "typing-extensions>=4.15.0",
    "unstructured>=0.18.15",
]

[project.scripts]
kickoff = "rag_flow.main:kickoff"
run_crew = "rag_flow.main:kickoff"
plot = "rag_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
